==> 2 Analysis/Proteome/Combined/3 Calculate relative amplitude.Rmd <==
---
title: "Estimating relative amplitude"
output: 
---

```{r}
#######################################################################################################################################
# Combined analysis Notebook 3 - DESCRIPTION
#
# Here we estimate relative amplitude for each protein.
# Relative amplitude = range / mean abundance,
# To calculate range, we use a 24-hour moving average method to detrend data before subtracting max - min.
# This allows the elimination of >24h trends (e.g. if everything is decreasing over several days)
#
# A relative amplitude threshold is used in proteomics analysis to define biological significance
# This threshold is chosen in this study to be 10%
#
# INPUTS = normalised data for significantly rhythmic proteins output by RAIN (.csv)
# OUTPUTS = detrended data, lists of proteins with relative amplitudes and baselines (mean abundance) calculated (.csv)
# 
#######################################################################################################################################
```

```{r}
# Setup
# Clear your workspace
rm(list=ls())

# Load useful packages
if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
} 

if (!require("gridExtra")) {
  install.packages("gridExtra", dependencies = TRUE)
  library(gridExtra)
}

# Make a list of all the .csv files in your working directory just so you can see what's going on
X = list.files(".",".csv")

# Read the raw data files
WT_data <- read.csv("WT_proteome_all_significant_RAIN.csv")
WT_data_start <- WT_data[,-1:-7]
WT_data_start <- t(WT_data_start)
colnames(WT_data_start) <- WT_data$Uniprot_ID


CKO_data <- read.csv("CKO_proteome_all_significant_RAIN.csv")
CKO_data_start <- CKO_data[,-1:-7]
CKO_data_start <- t(CKO_data_start)
colnames(CKO_data_start) <- CKO_data$Uniprot_ID

```

```{r}
# Detrend the data using a 24h moving average
# Set up useful parameters for detrending

period = 24     # Set period length to detrend
n_samples = ncol(WT_data_start)   # Total number of samples

time <- seq(24,93,3)   # List of time-points
row_days = round(period/(time[2] - time [1]))    # Number of rows that correspond to 24 hours (or 1 period length)

WT_detrended_data <- matrix(nrow=length(WT_data_start[,1]) - (row_days), ncol=ncol(WT_data_start))  #initialise a new matrix for the detrended data, excluding the first and last 12 hours (i.e. half of period)
colnames(WT_detrended_data) <- colnames(WT_data_start)     #set up the column names

#Fill up the matrix with the detrended data
for (i in 1:n_samples){
  for (j in 1:length(WT_detrended_data[,1])){
   WT_detrended_data[j,i] = WT_data_start[row_days/2+j,i] - mean(WT_data_start[j:(j + row_days),i])
  }
}

#Export the detrended data as a .csv file
write.csv(WT_detrended_data, file = "WT_Detrended_data.csv", row.names = F)


# SAME FOR CKO
period = 24     #set period length to detrend
n_samples = ncol(CKO_data_start)   #total number of samples
time <- seq(24,93,3)   #list of time-points
row_days = round(period/(time[2] - time [1]))    #number of rows that correspond to 24 hours (or 1 period length)

CKO_detrended_data <- matrix(nrow=length(CKO_data_start[,1]) - (row_days), ncol=ncol(CKO_data_start))  #initialise a new matrix for the detrended data, excluding the first and last 12 hours (i.e. half of period)
colnames(CKO_detrended_data) <- colnames(CKO_data_start)     #set up the column names

#Fill up the matrix with the detrended data
for (i in 1:n_samples){
  for (j in 1:length(CKO_detrended_data[,1])){
   CKO_detrended_data[j,i] = CKO_data_start[row_days/2+j,i] - mean(CKO_data_start[j:(j + row_days),i])
  }
}

#Export the detrended data as a .csv file
write.csv(CKO_detrended_data, file = "CKO_Detrended_data.csv", row.names = F)
```

```{r}
#Calculate relative amplitude

#Calculate Means of each column
WT_Baselines <- as_tibble(colMeans(WT_data_start))
WT_Baselines <- rename(WT_Baselines, Means = value)

#Calculate peak-trough amplitude from detrended data

WT_Amplitudes <- as_tibble(apply(WT_detrended_data, 2, function(df) {max(df)-min(df)}))

WT_Relative_amplitudes <- as_tibble(WT_Amplitudes/WT_Baselines)
WT_Relative_amplitudes <- rename(WT_Relative_amplitudes, WT_Relative_amplitudes = value)


# SAME FOR CKO
#Calculate Means of each column
CKO_Baselines <- as_tibble(colMeans(CKO_data_start))
CKO_Baselines <- rename(CKO_Baselines, Means = value)

#Calculate peak-trough amplitude from detrended data

CKO_Amplitudes <- as_tibble(apply(CKO_detrended_data, 2, function(df) {max(df)-min(df)}))

CKO_Relative_amplitudes <- as_tibble(CKO_Amplitudes/CKO_Baselines)
CKO_Relative_amplitudes <- rename(CKO_Relative_amplitudes, CKO_Relative_amplitudes = value)
```

```{r}
#Export a file with only those with relative amplitude ≥0.1, as well as files with complete data

WT_total_data <- cbind(WT_data[1:4], WT_Relative_amplitudes, WT_data[5:ncol(WT_data)])

WT_big_hits <- WT_total_data[WT_total_data$WT_Relative_amplitudes >= 0.1,]

write.csv(WT_total_data, "WT_all_hits.csv", row.names = F)
write.csv(WT_big_hits, "WT_big_hits.csv", row.names = F)



CKO_total_data <- cbind(CKO_data[1:4], CKO_Relative_amplitudes, CKO_data[5:ncol(CKO_data)])

CKO_big_hits <- CKO_total_data[CKO_total_data$CKO_Relative_amplitudes >= 0.1,]
write.csv(CKO_total_data, "CKO_all_hits.csv", row.names = F)
write.csv(CKO_big_hits, "CKO_big_hits.csv", row.names = F)
```


```{r}
# Export files that include baselines - this was added on later for different analyses

WT_export <- cbind(WT_data[1:4], WT_Relative_amplitudes, WT_Baselines, WT_data[5:ncol(WT_data)])
CKO_export <- cbind(CKO_data[1:4], CKO_Relative_amplitudes, CKO_Baselines, CKO_data[5:ncol(WT_data)])

write.csv(WT_export, "WT_RA_B.csv", row.names = F)
write.csv(CKO_export, "CKO_RA_B.csv", row.names = F)

```

```{r}
# Compare baselines of WT vs CKO rhythmic proteins
# t tests are fine as long as you use log-transformed data

WT_export$logbase <- log10(WT_export$Means)
CKO_export$logbase <- log10(CKO_export$Means)

test <- t.test(WT_export$logbase, CKO_export$logbase)

```

```{r}
# Compare the baselines of all WT + CKO proteins (including non-rhythmic)

WT_total <- read.csv("WT_RAIN_results.csv", stringsAsFactors = F)
CKO_total <- read.csv("CKO_RAIN_results.csv", stringsAsFactors = F)

WT_total$baselines <- apply(WT_total[,8:ncol(WT_total)], 1, mean)
WT_total$logbase <- log10(WT_total$baselines)

CKO_total$baselines <- apply(CKO_total[,8:ncol(CKO_total)], 1, mean)
CKO_total$logbase <- log10(CKO_total$baselines)

test1 <- t.test(WT_total$logbase, CKO_total$logbase)

```

```{r}
# Comparing rhythmic vs non-rhythmic in both genotypes --> include proteins with RA<10% because the question is about whether increasing abundance increases chance of detecting rhythmicity

WT_non <- read.csv("WT_proteome_all_insignificant_RAIN.csv", stringsAsFactors = F)
CKO_non <- read.csv("CKO_proteome_all_insignificant_RAIN.csv", stringsAsFactors = F)

WT_non$baselines <- apply(WT_non[,8:ncol(WT_non)], 1, mean)
WT_non$logbase <- log10(WT_non$baselines)

CKO_non$baselines <- apply(CKO_non[,8:ncol(CKO_non)], 1, mean)
CKO_non$logbase <- log10(CKO_non$baselines)


WT_export$logbase <- log10(WT_export$Means)
CKO_export$logbase <- log10(CKO_export$Means)

test3 <- t.test(WT_non$logbase, WT_export$logbase)

test4 <- t.test(CKO_non$logbase, CKO_export$logbase)

```

```{r}
# Comparing proteins not rhythmic in WT that become rhythmic in CKO, and the converse

CKO_new <- CKO_export[which(CKO_export$Uniprot_ID %in% WT_non$Uniprot_ID),]
CKO_new$WT_logbase <- WT_non$logbase[which(WT_non$Uniprot_ID %in% CKO_new$Uniprot_ID)]
CKO_new$WT_means <- WT_non$baselines[which(WT_non$Uniprot_ID %in% CKO_new$Uniprot_ID)]

test5 <- t.test(CKO_new$logbase, CKO_new$WT_logbase, paired = T)


WT_new <- WT_export[which(WT_export$Uniprot_ID %in% CKO_non$Uniprot_ID),]
WT_new$CKO_logbase <- CKO_non$logbase[which(CKO_non$Uniprot_ID %in% WT_new$Uniprot_ID)]
WT_new$CKO_means <- CKO_non$baselines[which(CKO_non$Uniprot_ID %in% WT_new$Uniprot_ID)]

test6 <- t.test(WT_new$logbase, WT_new$CKO_logbase, paired = T)

# Export these lists of proteins - CKOnew = proteins rhythmic in CKO but not WT; WTnew = the converse.

write.csv(CKO_new, "CKOnew.csv", row.names = F)
write.csv(WT_new, "WTnew.csv", row.names = F)
```


```{r}
# Plot the distributions of baselines for sanity-check - comparing rhythmic vs non-rhythmic
# NB: *_export refers to rhythmic proteins

axis_text_size = 12
axis_label_size = 12

histogram <- function(data) {
  ggplot(data, aes(x = logbase)) + 
  geom_histogram(binwidth = 0.1, col = "white") +
  xlim(3, 8) +
  theme(
  panel.border = element_blank(),  
  panel.grid.major = element_blank(),
  panel.grid.minor = element_blank(),
  panel.background = element_blank(),
  axis.line = element_line(colour = "black"),
  axis.text=element_text(colour="black", size = axis_text_size),
  axis.title = element_text(colour = "black", size = axis_label_size)) +
  labs(title = deparse(substitute(data))) +
  ylab("Count")
}


p1 <- histogram(WT_export)
p2 <- histogram(WT_non)
p3 <- histogram(CKO_export)
p4 <- histogram(CKO_non)

grid.arrange(p1, p3, p2, p4, nrow = 2)

```
