==> 2 Analysis/Proteome/Combined/5 Volcano plotter.Rmd <==
---
title: "Volcano plotting"
output: 
---

```{r}
#######################################################################################################################################
# Combined analysis Notebook 5 - DESCRIPTION
#
# Here we make volcano plots.
# In particular, we highlight proteins of interest by configuring the EnhancedVolcano function.
# Note that it is able to take some standard ggplot commands.
# Some of our configurations are not documented by the official documentation - these were discovered by blood, sweat and tears!
# Dual colours are also possible, but usually they look too busy. Connecting lines also make it too busy.
#
# INPUTS = results of mutliple t tests with BH correction (.csv)
# OUTPUTS = volcano plots (.pdf)
# 
#######################################################################################################################################
```

```{r}
# Setup
# Clear your workspace
rm(list=ls())

# Load useful packages
if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
}

if (!require("BiocManager")) {
  install.packages("BiocManager", dependencies = TRUE)
}

if (!require("EnhancedVolcano")) {
  BiocManager::install("EnhancedVolcano")
  library(EnhancedVolcano)
} 

if (!require("gridExtra")) {
  install.packages("gridExtra", dependencies = TRUE)
  library(gridExtra)
}

# Make a list of all the .csv files in your working directory just so you can see what's going on
X = list.files(".",".csv")

# Read the raw data files

results <- read.csv("CKOvsWT_results.csv", stringsAsFactors=FALSE)

```

```{r}
# First, plot a volcano to visualise the whole dataset

plt <- EnhancedVolcano(results,
    lab = as.vector(results$main_gene),
    x = 'log_fold_change',
    y = 'BH_pVal',
    ylab = bquote(~-Log[10]~italic(q)),
    axisLabSize = 14,
    title = NULL, # Title is optional
    subtitle = NULL,
    caption = NULL,
    xlim = c(min(results$log_fold_change), max(results$log_fold_change)), # Optimise axis limits
    ylim = c(0, max(-log10(results$BH_pVal))),
    pCutoff = 0.05,
    FCcutoff = log2(1), # not using a fold-change cutoff here!
    cutoffLineType = 'blank',
    cutoffLineCol = 'red',
    col = c("grey", "grey", "grey", "red"),
    shape = 16,
    pointSize = 4,
    labSize = 5,
    gridlines.major = FALSE,
    gridlines.minor = FALSE,
    legendPosition = "top",
    legendLabSize = 14,
    legendLabels=c(NA,'q>0.05', NA, 'q<=0.05'),
    legendIconSize= 5) +
    theme(panel.border = element_blank(), #make the annoying colour theme go away
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_blank(),
          axis.ticks = element_line(colour = "black"),
          axis.line = element_line(colour = "black"),
          axis.ticks.length=unit(0.1, "cm"),
          axis.text=element_text(colour="black", size = 12, family = "sans"), 
          legend.background = element_blank(),
          legend.key = element_blank()
          )

plt

ggsave("volcano_total_proteome.pdf", plot = plt, width=8, height = 9)

```


```{r}
# Import lists of gene names that you are interested in.
# These are created manually by going through the dataset and creating .csv files of your lists.
# Yes there are different importing methods - various reasons meant this was necessary to correctly read them.

rts <- read.csv("responsetostress.csv", stringsAsFactors = F, header = F)
rts <- rts$V1

```

```{r}
#######################################################################################################################################
# Creating a magnificent function for a single-colour volcano plot
#######################################################################################################################################

plt_volcano <- function(data, highlights) {
  # Reorder results with important ones first
  interest <- subset(data, data$main_gene %in% c(highlights))
  not_highlights <- subset(data, !(data$main_gene %in% c(highlights)))
  input <- rbind(not_highlights, interest)
  
  # Now make a named list to use as your instructions for plotting options
  keyvals.PQC <- rep('gray', nrow(input))
  names(keyvals.PQC) <- rep('Other', nrow(input))
  
  # Modify the set to highlight the ones you're interested in
  names(keyvals.PQC)[which(input$main_gene %in% highlights)] <- 'Highlights'
  keyvals.PQC[which(input$main_gene %in% highlights)] <- 'red'
  
  pd <- EnhancedVolcano(input,
    lab = as.vector(input$main_gene),
    selectLab = c(highlights), # This is to label only the ones you want
    x = 'log_fold_change',
    y = 'BH_pVal',
    ylab = bquote(~-Log[10]~italic(q)),
    axisLabSize = 14,
    title = NULL, # Title is optional
    subtitle = NULL,
    caption = NULL,
    xlim = c(min(input$log_fold_change), max(input$log_fold_change)), # Optimise laxis limits
    ylim = c(0, 50),
    pCutoff = 0.05,
    FCcutoff = log2(1), # not using a fold-change cutoff here!
    cutoffLineType = 'blank',
    cutoffLineCol = 'red',
    shape = 20,
    pointSize = c(ifelse(keyvals.PQC == 'gray', 4, 5)),
    labSize = 4,
    colCustom = keyvals.PQC,
    colAlpha = c(ifelse(keyvals.PQC == 'gray', 0.3, 1)),
    gridlines.major = FALSE,
    gridlines.minor = FALSE,
    legendPosition = "top",
    legendLabSize = 14,
    legendIconSize= 8) +
    theme(panel.border = element_blank(), #make the annoying colour theme go away
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_blank(),
          axis.ticks = element_line(colour = "black"),
          axis.line = element_line(colour = "black"),
          axis.ticks.length=unit(0.1, "cm"),
          axis.text=element_text(colour="black", size = 12), 
          legend.background = element_blank(),
          legend.key = element_blank()
          )
  pd
}
```

```{r}
# Do your plotting and export

# RESPONSE TO STRESS
plt <- plt_volcano(results, rts)
ggsave("volcano_RTS.pdf", plot = plt, width=5, height = 6)


```
